+++ extensions/oci/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Oracle Cloud Changelog
 
+## [] - {PR_MERGE_DATE}
+
+- Show a proper `Detail` view with information when "provider" fails. Previously, extension would crash (ref: Issue #).
+
 ## [Initial Version] - 2024-12-20
 
 - View Instances